Update package exports (#76)
To help be more explicit about how this module should be consumed, the `module` and `main` fields from `package.json` were removed. Since this module is meant to be run in a browser, we only keep the `browser` field and point it to our Common JS bundle. We still output the `scripts.js` and `scripts.min.js` bundles that can be used in a `<script>` tag.